Adding python to host for modin & modin-all.
These seem to be intended as metapackages that wrap the underlying backends. Since they are using exact pinning, we need one for each version of python supported. This means python needs to be in the host section. Another option would be to loosen the pinning to be at a version level: max_pin='x.x.x.x.x.x'